Translate action from DataDog to Signalfx
