+++ action.yml
@@ -1,21 +1,15 @@
-name: 'Datadog Action'
-description: 'Push Datadog metrics and events from GitHub Actions'
-author: 'Massimiliano Pippi'
+name: 'Signalfx Action'
+description: 'Push Signalfx events from GitHub Actions'
+author: 'Adapted by Falcon Taylor-Carter. Originaly created by Massimiliano Pippi'
 inputs:
-  api-key:
-    description: 'A valid Datadog API Key'
+  token:
+    description: 'A valid Signalfx token'
     required: true
   api-url:
-    description: 'The URL of the API endpoint, defaults to US datacenter'
-    default: 'https://api.datadoghq.com'
-  metrics:
-    description: 'A list of metrics objects to send, see docs for details'
-    default: '[]'
+    description: 'The URL of the ingest endpoint, defaults to US datacenter'
+    default: 'https://ingest.us1.signalfx.com'
   events:
     description: 'A list of event objects to send, see docs for details'
-    default: '[]'
-  service-checks:
-    description: 'A list of service check objects to send, see docs for details'
     default: '[]'
 runs:
   using: 'node12'
