+++ src/run.ts
@@ -1,25 +1,16 @@
 import * as core from '@actions/core'
-import * as dd from './datadog'
+import * as sfx from './signalfx'
 import * as yaml from 'js-yaml'
 
 export async function run(): Promise<void> {
   try {
-    const apiKey: string = core.getInput('api-key', {required: true})
+    const apiKey: string = core.getInput('token', {required: true})
     const apiURL: string =
-      core.getInput('api-url') || 'https://api.datadoghq.com'
+      core.getInput('api-url') || 'https://ingest.us1.signalfx.com'
 
-    const metrics: dd.Metric[] =
-      (yaml.safeLoad(core.getInput('metrics')) as dd.Metric[]) || []
-    await dd.sendMetrics(apiURL, apiKey, metrics)
-
-    const events: dd.Event[] =
-      (yaml.safeLoad(core.getInput('events')) as dd.Event[]) || []
-    await dd.sendEvents(apiURL, apiKey, events)
-
-    const serviceChecks: dd.ServiceCheck[] =
-      (yaml.safeLoad(core.getInput('service-checks')) as dd.ServiceCheck[]) ||
-      []
-    await dd.sendServiceChecks(apiURL, apiKey, serviceChecks)
+    const events: sfx.Event[] =
+      (yaml.safeLoad(core.getInput('events')) as sfx.Event[]) || []
+    await sfx.sendEvents(apiURL, apiKey, events)
   } catch (error) {
     core.setFailed(`Run failed: ${error.message}`)
   }
